Bug fix: log error and sink when parseError handler is undefined
If the client does not define a parseError handler (which is ok if
they are not using JSON for example) - then broken JSON parse flows
silently and only logs to __debug (which is mostly off in usual usage
and therefore never logged) and silently sinks the message.

Therefore, the client doesn't know if such an error was triggered
because handling parseError isn't forced (to maintain flexibility
where client wants to handle raw messages directly).

This change forces the message to be sinked and then emits a parseError
event. If not handler is defined, it will log an error so that the user
can deal with it.